// fe/src/lib/api/generated/model/identityUpdateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface IdentityUpdateRequest {
  first_name?: string | null;
  institutional_id?: string | null;
  last_name?: string | null;
  middle_name?: string | null;
  suffix?: string | null;
}